soc: imx: gpcv2: Add DISPMIX syscon interface for i.MX8MM
Replicate a rather nasty hack from TFA which programs the dispmix BLK_CTL
with undocumented values. This needs to be understood and cleaned up, but
for now keep it here as it is needed to get display working.

Also note that there is something wrong with clock, I suspect one of the
dispmix AHB clock need to be kept on to access the DSIM PHY registers
during probe, keep the disp clock running for now.

Upstream-Status: WIP
Signed-off-by: Marek Vasut <marex@denx.de>